+++ lib/cpalgo/algebra/combination_basic.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <stdexcept>
 #include <vector>
 
@@ -20,22 +21,21 @@
 //
 class Combination {
     size_t N;
-    unsigned long long MOD;
-    std::vector<unsigned long long> factorial, inverse, factorial_inverse;
+    uint64_t MOD;
+    std::vector<uint64_t> factorial, inverse, factorial_inverse;
 
 public:
     // O(N)
     // MOD should be a prime number
     Combination(
         size_t const N = 10000000,
-        unsigned long long const MOD = 1000000007
+        uint64_t const MOD = 1000000007
     )
     {
         build(N, MOD);
     }
-
     // Time: O(N)
-    void build(size_t const N, unsigned long long const MOD) {
+    void build(size_t const N, uint64_t const MOD) {
         this->N = N;
         this->MOD = MOD;
 
@@ -52,15 +52,13 @@ public:
             factorial_inverse[i] = factorial_inverse[i-1] * inverse[i] % MOD;
         }
     }
-
     // nCk
     // Time: O(1)
-    unsigned long long operator()(size_t const n, size_t const k) const {
+    uint64_t operator()(size_t const n, size_t const k) const {
         if (k > n) throw std::out_of_range("k");
         if (n >= N) throw std::out_of_range("n");
         return factorial[n] * (factorial_inverse[k] * factorial_inverse[n - k] % MOD) % MOD;
     }
-
     // Time: O(1)
     size_t size() const {
         return N;
