+++ lib/cpalgo/algebra/extgcd_iterative.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <tuple>
 
 
@@ -12,22 +13,21 @@
 //
 // Verified:
 //  - https://onlinejudge.u-aizu.ac.jp/courses/library/6/NTL/1/NTL_1_E
-//  - https://uva.onlinejudge.org/index.php?option=com_onlinejudge&Itemid=8&page=show_problem&problem=1045
 //
-inline unsigned long long extgcd(
-    unsigned long long const a,
-    unsigned long long const b,
-    long long &x,
-    long long &y
+inline uint64_t extgcd(
+    uint64_t const a,
+    uint64_t const b,
+    int64_t &x,
+    int64_t &y
 )
 {
     x = 1, y = 0;
-    long long x1 = 0, y1 = 1;
-    unsigned long long a1 = a, b1 = b;
+    int64_t x1 = 0, y1 = 1;
+    uint64_t a1 = a, b1 = b;
     // a * x  + b * y  = a1
     // a * x1 + b * y1 = b1
     while (b1 > 0) {
-        unsigned long long q = a1 / b1;
+        uint64_t q = a1 / b1;
         std::tie(x, x1) = std::make_tuple(x1, x - q * x1);
         std::tie(y, y1) = std::make_tuple(y1, y - q * y1);
         std::tie(a1, b1) = std::make_tuple(b1, a1 - q * b1);
