+++ lib/cpalgo/algebra/lcm.hpp
@@ -14,7 +14,7 @@
 //  - https://www.codechef.com/problems/FLOW016
 //  - https://onlinejudge.u-aizu.ac.jp/courses/library/6/NTL/1/NTL_1_C
 //
-inline unsigned long long lcm(unsigned long long const a, unsigned long long const b) {
+inline uint64_t lcm(uint64_t const a, uint64_t const b) {
     if (a == 0 && b == 0) return 0;
     return a / gcd(a, b) * b;
 }
